Move functionality to its own file
